Move network module
