+++ app/src/main/java/com/geeksville/mesh/repository/api/FirmwareReleaseLocalDataSource.kt
@@ -22,35 +22,28 @@ import com.geeksville.mesh.database.entity.FirmwareReleaseEntity
 import com.geeksville.mesh.database.entity.FirmwareReleaseType
 import com.geeksville.mesh.database.entity.asDeviceVersion
 import com.geeksville.mesh.database.entity.asEntity
-import com.geeksville.mesh.network.model.NetworkFirmwareRelease
 import dagger.Lazy
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+import org.meshtastic.core.network.model.NetworkFirmwareRelease
 import javax.inject.Inject
 
-class FirmwareReleaseLocalDataSource @Inject constructor(
-    private val firmwareReleaseDaoLazy: Lazy<FirmwareReleaseDao>
-) {
-    private val firmwareReleaseDao by lazy {
-        firmwareReleaseDaoLazy.get()
-    }
+class FirmwareReleaseLocalDataSource @Inject constructor(private val firmwareReleaseDaoLazy: Lazy<FirmwareReleaseDao>) {
+    private val firmwareReleaseDao by lazy { firmwareReleaseDaoLazy.get() }
 
     suspend fun insertFirmwareReleases(
         firmwareReleases: List<NetworkFirmwareRelease>,
-        releaseType: FirmwareReleaseType
-    ) =
-        withContext(Dispatchers.IO) {
-            if (firmwareReleases.isNotEmpty()) {
-                firmwareReleaseDao.deleteAll()
-                firmwareReleases.forEach { firmwareRelease ->
-                    firmwareReleaseDao.insert(firmwareRelease.asEntity(releaseType))
-                }
+        releaseType: FirmwareReleaseType,
+    ) = withContext(Dispatchers.IO) {
+        if (firmwareReleases.isNotEmpty()) {
+            firmwareReleaseDao.deleteAll()
+            firmwareReleases.forEach { firmwareRelease ->
+                firmwareReleaseDao.insert(firmwareRelease.asEntity(releaseType))
             }
         }
-
-    suspend fun deleteAllFirmwareReleases() = withContext(Dispatchers.IO) {
-        firmwareReleaseDao.deleteAll()
     }
+
+    suspend fun deleteAllFirmwareReleases() = withContext(Dispatchers.IO) { firmwareReleaseDao.deleteAll() }
 
     suspend fun getLatestRelease(releaseType: FirmwareReleaseType): FirmwareReleaseEntity? =
         withContext(Dispatchers.IO) {
@@ -58,8 +51,7 @@ class FirmwareReleaseLocalDataSource @Inject constructor(
             if (releases.isEmpty()) {
                 return@withContext null
             } else {
-                val latestRelease =
-                    releases.maxBy { it.asDeviceVersion() }
+                val latestRelease = releases.maxBy { it.asDeviceVersion() }
                 return@withContext latestRelease
             }
         }
