+++ core/network/src/main/kotlin/org/meshtastic/core/network/service/ApiService.kt
@@ -15,20 +15,16 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package com.geeksville.mesh.repository.api
+package org.meshtastic.core.network.service
 
-import android.app.Application
-import kotlinx.serialization.ExperimentalSerializationApi
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.json.decodeFromStream
+import de.jensklingenberg.ktorfit.http.GET
+import org.meshtastic.core.network.model.NetworkDeviceHardware
 import org.meshtastic.core.network.model.NetworkFirmwareReleases
-import javax.inject.Inject
 
-class FirmwareReleaseJsonDataSource @Inject constructor(private val application: Application) {
-    @OptIn(ExperimentalSerializationApi::class)
-    fun loadFirmwareReleaseFromJsonAsset(): NetworkFirmwareReleases {
-        val inputStream = application.assets.open("firmware_releases.json")
-        val result = inputStream.use { Json.decodeFromStream<NetworkFirmwareReleases>(inputStream) }
-        return result
-    }
+interface ApiService {
+    @GET("resource/deviceHardware")
+    suspend fun getDeviceHardware(): List<NetworkDeviceHardware>
+
+    @GET("github/firmware/list")
+    suspend fun getFirmwareReleases(): NetworkFirmwareReleases
 }
