--- network/build.gradle.kts
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2025 Meshtastic LLC
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-plugins {
-    alias(libs.plugins.meshtastic.android.library)
-    alias(libs.plugins.meshtastic.hilt)
-    alias(libs.plugins.kotlin.serialization)
-    alias(libs.plugins.dokka)
-    alias(libs.plugins.kover)
-    alias(libs.plugins.protobuf)
-    alias(libs.plugins.ktorfit)
-}
-
-android {
-    buildFeatures { buildConfig = true }
-    namespace = "com.geeksville.mesh.network"
-}
-
-dependencies {
-    implementation(libs.bundles.ktor)
-    implementation(libs.bundles.coil)
-    "googleImplementation"(libs.bundles.datadog)
-    implementation(libs.kotlinx.serialization.json)
-}
